+++ server/dlfaker.c
@@ -1,5 +1,6 @@
 /* Copyright (C)2006 Sun Microsystems, Inc.
- * Copyright (C)2009, 2012, 2015, 2017 D. R. Commander
+ * Copyright (C)2009, 2012, 2015, 2017-2018 D. R. Commander
+ * Copyright (C)2015 Open Text SA and/or Open Text ULC (in Canada)
  *
  * This library is free software and may be redistributed and/or modified under
  * the terms of the wxWindows Library License, Version 3.1 or (at your option)
@@ -16,10 +17,47 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <dlfcn.h>
+#include <pthread.h>
 #include "vendor.h"
 
 
 extern void *_vgl_dlopen(const char *, int);
+
+
+static pthread_key_t getDLFakerLevelKey(void)
+{
+	static pthread_key_t key;
+	static int init = 0;
+
+	if(!init)
+	{
+		if(pthread_key_create(&key, NULL))
+		{
+			fprintf(stderr, "[VGL] ERROR: pthread_key_create() for DLFakerLevel failed.\n");
+			return (pthread_key_t)-1;
+		}
+		pthread_setspecific(key, (const void *)0);
+		init = 1;
+	}
+	return key;
+}
+
+static int getDLFakerLevel(void)
+{
+	pthread_key_t dlFakerLevelKey = getDLFakerLevelKey();
+
+	if(dlFakerLevelKey == (pthread_key_t)-1)
+		return -1;
+	else return (int)(size_t)pthread_getspecific(dlFakerLevelKey);
+}
+
+static void setDLFakerLevel(int value)
+{
+	pthread_key_t dlFakerLevelKey = getDLFakerLevelKey();
+
+	if(dlFakerLevelKey != -1)
+		pthread_setspecific(dlFakerLevelKey, (const void *)(size_t)value);
+}
 
 
 /* If an application uses dlopen()/dlsym() to load functions from libGL or
@@ -37,6 +75,15 @@ void *dlopen(const char *filename, int flag)
 	char *env = NULL, *env2 = NULL;  const char *envname = "FAKERLIB32";
 	int verbose = 0, trace = 0;
 	void *retval = NULL;
+
+	/* Prevent segfault if an application interposes getenv() and calls dlopen()
+	   within the body of its interposed getenv() function (I'm looking at you,
+	   Cadence Virtuoso)
+	*/
+	if(getDLFakerLevel() > 0)
+		return _vgl_dlopen(filename, flag);
+
+	setDLFakerLevel(getDLFakerLevel() + 1);
 
 	if(sizeof(long) == 8) envname = "FAKERLIB";
 	if((env2 = getenv("VGL_VERBOSE")) != NULL && strlen(env2) > 0
@@ -95,6 +142,8 @@ void *dlopen(const char *filename, int flag)
 	{
 		fprintf(stderr, " retval=0x%.8lx)\n", (long)retval);
 	}
+
+	setDLFakerLevel(getDLFakerLevel() - 1);
 
 	return retval;
 }